pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fastweb3"
version = "0.1.1"
description = "High-performance Web3 client with batching and latency optimization."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
    { name = "iamdefinitelyahuman" }
]

dependencies = [
    "httpx>=0.27",
    "lazy-object-proxy>=1.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-cov>=5",
    "ruff>=0.9",
    "pre-commit>=3",
    "build",
    "twine",
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
    "websocket-client>=1.6",
]
docs = [
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
]
wss = ["websocket-client>=1.6"]

[project.urls]
Homepage = "https://github.com/iamdefinitelyahuman/fastweb3"
Repository = "https://github.com/iamdefinitelyahuman/fastweb3"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
addopts = [
  "--cov=fw3",
  "--cov-report=term-missing",
  "--cov-report=xml",
]
testpaths = ["tests"]